requirements.txt
alembic==1.4.3
aniso8601==8.0.0
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
mysqlclient==2.0.2
pkg-resources==0.0.0
protobuf==3.14.0
pycparser==2.20
PyJWT==2.0.1
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
Werkzeug==1.0.1
